requirements.txt
asgiref==3.3.1
coverage==5.3.1
Django==3.1.4
django-jazzmin==2.4.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
flake8==3.8.4
mccabe==0.6.1
Pillow==8.0.1
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==2.0.0
pytz==2020.4
sqlparse==0.4.1
